Hold the viewer scene back until the simulator work
`scene.xml` is not from upstream: it puts the arm on a lit ground plane for a
viewer to open. Nothing in the code reads it. The vendored model stays, since
the driver solves its kinematics against it, and its README now states the
layout as it is and names the revision it came from.
