--- positronic/assets/mujoco/trossen_wxai/scene.xml
@@ -1,24 +1,0 @@
-<mujoco model="wxai follower scene">
-  <include file="wxai_follower.xml"/>
-
-  <statistic center="0.3 0 0.2" extent="0.9"/>
-
-  <visual>
-    <headlight diffuse="0.8 0.8 0.8" ambient="0.45 0.45 0.45" specular="0.2 0.2 0.2"/>
-    <rgba haze="0.15 0.25 0.35 1"/>
-    <global azimuth="150" elevation="-20"/>
-  </visual>
-
-  <asset>
-    <texture type="skybox" builtin="gradient" rgb1="0.3 0.5 0.7" rgb2="0 0 0" width="512" height="3072"/>
-    <texture type="2d" name="groundplane" builtin="checker" mark="edge" rgb1="0.35 0.45 0.55" rgb2="0.25 0.35 0.45"
-      markrgb="0.9 0.9 0.9" width="300" height="300"/>
-    <material name="groundplane" texture="groundplane" texuniform="true" texrepeat="5 5" reflectance="0.2"/>
-  </asset>
-
-  <worldbody>
-    <light pos="0.6 0.6 1.2" dir="-0.4 -0.4 -1" directional="true" diffuse="0.7 0.7 0.7"/>
-    <light pos="-0.4 -0.6 1.0" dir="0.3 0.5 -1" directional="true" diffuse="0.5 0.5 0.5"/>
-    <geom name="floor" size="0 0 0.05" type="plane" material="groundplane" pos="0 0 -0.01"/>
-  </worldbody>
-</mujoco>
